+++ ml_model/train.py
@@ -20,7 +20,7 @@ from sklearn.metrics import (
     classification_report,
     confusion_matrix,
 )
-from processing import data_wrangler
+from preprocessing import data_wrangler
 warnings.filterwarnings('ignore')
